task

+++ 0x09-static_libraries/2-strlen.c
@@ -1,5 +1,4 @@
 #include "main.h"
-#include <stddef.h>
 
 
 /**
@@ -11,7 +10,7 @@
 int _strlen(char *str)
 {
 
-	size_t length = 0;
+	int length = 0;
 
 	while (*str++)
 		length++;
